requirements.txt
tensorflow-cpu
fastapi
uvicorn
Pillow
numpy